First Checkin
First check in of the Arduino sketches for the Command Center Beta project.
- ButtonTest Sketch to test the Controller Buttons
- JoystickPadTest Sketch to test the Controller Joystick
- DefaultController Sketch to be used in the tutorials.

The bulk of the code is in the DefaultController. In this iteration, the controller
functions as a mouse in one mode. When the joystick button is pressed it
functions it will function as a keyboard controller, with the joystick being the
corresponding arrow keys, and the buttons being mapped to key presses found
in the code.